+++ browser/data-browser/src/routes/History/useVersions.ts
@@ -1,32 +1,48 @@
-import { Resource, Version, useStore } from '@tomic/react';
-import { useState, useEffect } from 'react';
+import { Resource, Version, unknownSubject, useStore } from '@tomic/react';
+import { useState, useEffect, useRef, useTransition } from 'react';
 import { dedupeVersions } from './versionHelpers';
 
 export interface UseVersionsResult {
   versions: Version[];
   loading: boolean;
+  progress: number;
   error: Error | undefined;
 }
 
 export function useVersions(resource: Resource): UseVersionsResult {
   const [versions, setVersions] = useState<Version[]>([]);
+  const [progress, setProgress] = useState(0);
+  const isRunning = useRef(false);
   const store = useStore();
+  const [_, startTransition] = useTransition();
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<Error | undefined>(undefined);
 
   useEffect(() => {
-    resource
-      .getHistory(store)
-      .then(history => {
-        setVersions(dedupeVersions(history));
-      })
-      .catch(e => {
-        setError(e);
-      })
-      .finally(() => {
-        setLoading(false);
-      });
+    if (resource.getSubject() === unknownSubject) {
+      return;
+    }
+
+    if (isRunning.current) {
+      return;
+    }
+
+    startTransition(() => {
+      (async () => {
+        try {
+          isRunning.current = true;
+          const history = await resource.getHistory(store, setProgress);
+          const dedupedVersions = dedupeVersions(history);
+          setVersions(dedupedVersions);
+        } catch (e) {
+          setError(e);
+        } finally {
+          setLoading(false);
+          isRunning.current = false;
+        }
+      })();
+    });
   }, [resource]);
 
-  return { versions, loading, error };
+  return { versions, loading, error, progress };
 }