+++ browser/CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog covers all three packages, as they are (for now) updated as a who
 ### @tomic/lib
 
 - Always fetch all resources after setting + authenticating new agent with websockets #686
+- Add progress callback to `resource.getHistory()` And increased its performance for resources with a large number of commits [#745](https://github.com/atomicdata-dev/atomic-server/issues/745)
 
 ## v0.36.1